Refactor node bootstrapping into own component
