+++ website/docs/tutorials/queries.md
@@ -32,10 +32,28 @@ last_studied = { type = "int" }
 ```
 
 ```jsx live
+function BootstrapNodeApp(props) {
+  const VOCABULARY_SCHEMA_ID =
+    'vocabulary_0020a14f78ab4950661db6e5a44849c3770735cb723abc08097d65bbfe0ad8ddcd11';
+  const ENDPOINT = 'http://wolke.liebechaos.org:2020';
+
+  return (
+    <App header="🐬 🆗" endpoint={ENDPOINT}>
+      <BootstrapNode
+        schemaLockUrl="/schemas/query-tutorial/schema.lock"
+        schemaId={VOCABULARY_SCHEMA_ID}
+      ></BootstrapNode>
+    </App>
+  );
+}
+```
+
+```jsx live
 function QueriesApp(props) {
   const VOCABULARY_SCHEMA_ID =
     'vocabulary_0020a14f78ab4950661db6e5a44849c3770735cb723abc08097d65bbfe0ad8ddcd11';
-
+  const ENDPOINT = 'http://wolke.liebechaos.org:2020';
+  
   const query = `
     all_${VOCABULARY_SCHEMA_ID}(filter: { word: { contains: "狭" }}) {
       totalCount
@@ -49,7 +67,7 @@ function QueriesApp(props) {
   `;
 
   return (
-    <App header="&#128122; &#128218;" endpoint="http://wolke.liebechaos.org:2020">
+    <App header="👺 📚" endpoint={ENDPOINT}>
       <QueryTutorial query={query}></QueryTutorial>
     </App>
   );
