+++ website/src/theme/ReactLiveScope/P2pandaContext.jsx
@@ -4,7 +4,7 @@ import { GraphQLClient } from 'graphql-request';
 
 const LOCAL_STORAGE_KEY = 'privateKey';
 
-const ENDPOINT = 'http://localhost:2020/graphql';
+const ENDPOINT = 'http://localhost:2020';
 
 function getKeyPair() {
   // eslint-disable-next-line no-undef
@@ -26,11 +26,11 @@ export const P2pandaContext = React.createContext({
   graphQLClient: null,
 });
 
-export const P2pandaProvider = ({ children }) => {
+export const P2pandaProvider = ({ children, endpoint = ENDPOINT }) => {
   const state = useMemo(() => {
     const keyPair = getKeyPair();
-    const session = new Session(ENDPOINT).setKeyPair(keyPair);
-    const graphQLClient = new GraphQLClient(ENDPOINT);
+    const session = new Session(`${endpoint}/graphql`).setKeyPair(keyPair);
+    const graphQLClient = new GraphQLClient(`${endpoint}/graphql`);
 
     return {
       keyPair,
@@ -38,7 +38,7 @@ export const P2pandaProvider = ({ children }) => {
       session,
       graphQLClient,
     };
-  }, []);
+  }, [endpoint]);
 
   return (
     <P2pandaContext.Provider value={state}>{children}</P2pandaContext.Provider>
