+++ website/src/theme/ReactLiveScope/index.jsx
@@ -40,11 +40,24 @@ export function QueryTutorial(props) {
   );
 }
 
+export function BootstrapNode(props) {
+  return (
+    <BrowserOnly fallback={<div>...</div>}>
+      {() => {
+        // eslint-disable-next-line @typescript-eslint/no-var-requires, no-undef
+        const Component = require('./BootstrapNode').BootstrapNode;
+        return <Component {...props} />;
+      }}
+    </BrowserOnly>
+  );
+}
+
 // Add react-live imports you need here
 const ReactLiveScope = {
   React,
   ...React,
   App,
+  BootstrapNode,
   CafeForm,
   QueryTutorial,
 };
